requirements.txt

speedtest-cli
future
faker
motor
PyDictionary==2.0.1
emoji
lxml
beautifulsoup4
requests
sqlalchemy==1.3.20
python-telegram-bot==13.9
psycopg2-binary
feedparser
pynewtonmath
spongemock
zalgo-text
geopy
psutil
jikanpy
coffeehouse
regex
bleach
wikipedia
telethon
spamwatch
markdown2
alphabet_detector
cachetools
fontTools
gtts
tswift
pytz
PyYAML>=5.3.1
redis
html2text
pythonping
html_telegraph_poster
pytube
youtube-search-python
instantmusic
PyLyrics
hachoir
pybase64
pySmartDL
validators
pyrogram
nltk
aiohttp>=3.7.4
pykeyboard==0.1.2
Pillow>=4.2.0
CurrencyConverter
googletrans==4.0.0-rc1
google_trans_new
pyrate-limiter
nekos.py
tgcrypto
Python_ARQ
hachoir
configparser
bing_image_downloader
parsel
urllib3
heroku3
telegraph
youtube_search
search_engine_parser
glitch_this
NumPy
opencv-python-headless
countryinfo==0.1.2
secureme #encrypt
aiofiles
wget
tracemoepy #whatanime.py
img2pdf
gpytranslate
pretty_errors
psutil
